requirements.txt
absl-py==0.7.1
astor==0.8.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
gast==0.2.2
grpcio==1.20.1
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.0.0
protobuf==3.7.1
requests==2.22.0
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
urllib3==1.25.2
Werkzeug==0.15.4
